Updated VisualStudio project and source to remove warnings
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/RobWork/trunk@1258 494c09c7-a67d-49d4-9938-151fb1965be8

+++ src/rw/common/Log.cpp
@@ -123,7 +123,7 @@ void Log::remove(LogLevel id)
 }
 
 bool Log::isValidLogLevel(LogLevel id){
-	if(id<0 || _writers.size()<id)
+	if(id<0 || _writers.size()<(size_t)id)
 		return false;
 	if(_writers[id]==NULL)
 		return false;
